+++ entity_event/tests/models_tests.py
@@ -8,7 +8,7 @@ from freezegun import freeze_time
 from six import text_type
 
 from entity_event.models import (
-    Medium, Source, SourceGroup, Unsubscription, Subscription, Event, EventActor, EventSeen
+    Medium, Source, SourceGroup, Unsubscription, Subscription, Event, EventActor, EventSeen, _unseen_event_ids
 )
 
 
@@ -225,6 +225,27 @@ class MediumSubsetSubscriptionsTest(TestCase):
         self.assertEqual(subs.count(), 0)
 
 
+class MediumGetFilteredEventsTest(TestCase):
+    def setUp(self):
+        self.medium = G(Medium)
+
+    def test_get_unseen_events_some_seen_some_not(self):
+        seen_e = G(Event, context={})
+        G(EventSeen, event=seen_e, medium=self.medium)
+        unseen_e = G(Event, context={})
+
+        events = self.medium.get_filtered_events(seen=False)
+        self.assertEquals(list(events), [unseen_e])
+
+    def test_get_unseen_events_some_seen_from_other_mediums(self):
+        seen_from_other_medium_e = G(Event, context={})
+        G(EventSeen, event=seen_from_other_medium_e)
+        unseen_e = G(Event, context={})
+
+        events = self.medium.get_filtered_events(seen=False)
+        self.assertEquals(set(events), set([unseen_e, seen_from_other_medium_e]))
+
+
 class MediumGetEventFiltersTest(TestCase):
     def setUp(self):
         self.medium = G(Medium)
@@ -365,6 +386,28 @@ class SubscriptionSubscribedEntitiesTest(TestCase):
     def test_length_indiv(self):
         indiv_qs = self.indiv_sub.subscribed_entities()
         self.assertEqual(indiv_qs.count(), 1)
+
+
+class UnseenEventIdsTest(TestCase):
+    def test_filters_seen(self):
+        m = G(Medium)
+        e1 = G(Event, context={})
+        e2 = G(Event, context={})
+        Event.objects.filter(id=e2.id).mark_seen(m)
+        unseen_ids = _unseen_event_ids(m)
+        self.assertEqual(unseen_ids, [e1.id])
+
+    def test_multiple_mediums(self):
+        m1 = G(Medium)
+        m2 = G(Medium)
+        e1 = G(Event, context={})
+        e2 = G(Event, context={})
+        e3 = G(Event, context={})
+        e4 = G(Event, context={})
+        Event.objects.filter(id=e2.id).mark_seen(m1)
+        Event.objects.filter(id=e3.id).mark_seen(m2)
+        unseen_ids = _unseen_event_ids(m1)
+        self.assertEqual(set(unseen_ids), set([e1.id, e3.id, e4.id]))
 
 
 # Note: The following freeze_time a few more minutes than what we