+++ docs/release_notes.rst
@@ -1,6 +1,16 @@
 Release Notes
 =============
 
+v0.2
+----
+
+* This release provides the core features of django-entity-event
+  - Event Creation
+  - Subscription Management
+  - Event Querying
+  - Admin Panel
+  - Documentation
+
 v0.1
 ----